requirements.txt
matplotlib-inline==0.1.3
sklearn==0.0
setuptools==58.2.0
pip==21.3.1
Flask==2.2.2
render_template==2.0.1
request==2.26.0 
jsonify==2.2.0
CORS==3.0.10 
DecisionTreeRegressor==0.24.2 
matplotlib.pyplot==3.4.3
numpy==1.21.3
base64==1.5
BytesIO==0.2.1 